requirements.txt
openai
python-dotenv
streamlit
gtts
pygame